moving interfaces, separating reducer functions
