+++ src/app/Breadcrumb.tsx
@@ -4,7 +4,7 @@ import { CollapsedBreadcrumb } from '../components/breadcrumb/CollapsedBreadcrum
 import { SingleBreadcrumb } from '../components/breadcrumb/SingleBreadcrumb';
 import { useAppDispatch } from '../hooks/useAppDispatch';
 import { useAppSelector } from '../hooks/useAppSelector';
-import { changeFocus, setCreateNextWorkbenchSidebarOpen } from '../store/ordinoSlice';
+import { changeFocus, setCreateNextWorkbenchSidebarOpen } from '../store';
 
 export const colorPalette = ['#337ab7', '#ec6836', '#75c4c2', '#e9d36c', '#24b466', '#e891ae', '#db933c', '#b08aa6', '#8a6044', '#7b7b7b'];
