+++ src/app/workbench/sidebar/DetailsSidebar.tsx
@@ -3,7 +3,7 @@ import { IDType, useAsync, ViewUtils } from 'tdp_core';
 import { IReprovisynMapping } from 'reprovisyn';
 import { useAppSelector } from '../../../hooks/useAppSelector';
 import { useAppDispatch } from '../../../hooks/useAppDispatch';
-import { changeSelectedMappings, IWorkbench } from '../../../store/ordinoSlice';
+import { changeSelectedMappings, IWorkbench } from '../../../store';
 import { isVisynRankingViewDesc } from '../../../views/interfaces';
 
 export interface IDetailsSidebarProps {
