+++ src/store/index.ts
@@ -1,4 +1,5 @@
 export * from './appSlice';
+export * from './interfaces';
 export * from './ordinoSlice';
 export * from './menuSlice';
 export * from './store';
